Change leading \s to [[:space:]].
As found in basho/riak_pb#94, on FreeBSD 9, the "\s" shortcut
character class leads to not filtering out the desired lines, meaning
we were left with a non-empty dialyzer_unhandled_warnings, even if all
known warnings were filtered by the second grep. Changing this to
[[:space:]] resolves the issue, and is more consistent with the
previous 'grep' call in the step.