Ignore line numbers in dialyzer ignore file
This change uses sed to remove the line numbers from the ignore file and
the warnings file. That way the comparison doesn't care if line numbers
change, which happens very often.

Notice that the final output does not have the current line number for
the warning. I think that is OK for now.  With some trickery involving
sort and sed the original line numbers could be re-added to the output,
but that will have to wait for another day.