Show original line numbers even when ignored
A previous set of changes now makes tools.mk ignore line numbers when
matching ignored warnings, even though it looks at the number of equal
warnings to make sure and alert the user if we have more than expected.
One side effect of that change is that the line numbers are removed from
the output, which is confusing.
This change improves on that by showing the line numbers from the
original warnings output. It may match some warnings that are already
ignored, but that should be harmless mostly. If anything, it will make
people look at those ignored lines again.

This also fixes a problem with that previous code. Sorting was happening
before the line numbers were removed, so some duplicates may not end up
grouped together.