Update PHONY targets
Specifically, the `docs` target would fail to run in any repository that
had a `docs/` directory, like `riak-erlang-client`. This is because the
target was always considered up-to-date. You can test this by running
`make docs -n` with `riak-erlang-client` before and after this tools.mk
change. None of the other PHONY targets are likely to have
file/directory collisions, but it's better to add them just in case.